Add custom exception message for fatal error
Modified the fatal exception raised by connect() to ensure graceful
exit. After a fixed number of retries, the connect() method now raises
a fatal exception with a custom message instead of using the string
representation of the last socket error.
This is to ensure compatibility with Python 3 (see PEP 3110 for changes
to exceptions). In Python 3, targets assigned to exceptions in the
except clause are explicitly deleted after the except block is exited.
Caught exceptions are not available for use after the except block
unless they are reassigned to a variable in the block.
Using the socket error for raising the fatal exception actually raises
an UnboundLocalError with Python 3 here, instead of raising
FatalTraCIError and exiting gracefully.