Fixed a copy command (re #3836)
The build error in #3836 is caused by the command `cp`, which supposedly only works on Windows when cygwin is installed . The usual command to copy files on Windows is `copy`. However, `copy` does not work properly when paths contain slashes. I therefore replaced the copy command with the CMake built-in copy command. I also added `$<CONFIG>` to the path, as otherwise the command will fail if Sumo isn't compiled in Release mode.

For the same reason, I added the path to the Debug build in the following `FIND_LIBRARY`call.