Add router dependencies to custom examples target
The tests listed in tests/examples.test require that the executables of dfrouter, duarouter and jtrrouter are available when the python scripts are run. When they are not listed as dependencies to the custom target, the build of the examples target might fail, depending on which targets are build first. Adding the *router targets as dependencies fixes that issue.
Sidenote: Building the examples target in Debug always fails if no Release build has been made previously, as the names of all debug executables are suffixed with "D" and thus are not found/used by the python scripts.