Improved platoonmanager
The improvements are these:

1. before a vehicle starts following a leader, it checks if at least one of these conditions are satisfied:

- leader and ego vehicle have a certain number of edges in common on the future route

- leader and ego vehicle do a certain distance together on a future route

2. The lane change advice is not issued to a follower if it is less than a certain distance from the upstream intersection


The former feature reduces the number of unnecessary catchups and the whole simulation becomes quieter and less chaotic. I have not yet performance figure though.

The latter prevents vehicles to be pushed on a lane with dead end (or occupied) short before the intersection. This measure prevents vehicles getting stuck, at least for the default lanechange model. With the sublane model it can still happen that vehicles get stuck in roundabouts. Increasing the strategy factor and pushy factor helps to improve this situation, but does not resolve it completely.

Parameters are set in __init__():
 # ego vehicle needs at least this number of future edges in common with leader before agreeing to follow...
        self._edgenumberLookahead = 3

        # Or the ego vehicle needs at least this distance of common route length with leader before agreeing to follow.
        self._distLookahead = 500.0

        # no lane change advice if vehicle has less than this distance to the next  junction
        self._lanechangeMinDist = 100.0

Touched only functions:
_updateVehicleStates(self):
_adviseLanes(self):